pyproject.toml
[project]
name = "sole"
version = "0.1.0"
description = "A tool for dynamically generating markdown files using references to commands."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Solenya AI", email = "Solenya-AIaaS@users.noreply.github.com" }]
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
]
keywords = ["documentation", "markdown", "generator", "sole", "devops"]
urls = { Repository = "https://github.com/Solenya-AIaaS/sole", Issues = "https://github.com/Solenya-AIaaS/sole/issues" }

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "vulture",
    "complexipy",
    "ruff",
]

[project.scripts]
sole = "sole.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sole"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
    # pydoclint
    "DOC",
    # mccabe
    "C90",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.vulture]
min_confidence = 80
paths = ["src", "tests"]
sort_by_size = true

[tool.complexipy]
max_complexity = 15

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
